Change `removeSearch` logic to run getSearchHistory() in all situations

+++ app/src/main/java/cc/wordview/app/ui/activities/home/composables/search/SearchViewModel.kt
@@ -177,9 +177,11 @@ class SearchViewModel @Inject constructor(
     }
 
     fun removeSearch(searchQuery: String) = viewModelScope.launch(Dispatchers.IO) {
-        val search = searchQueryDao.findByQuery(searchQuery) ?: return@launch
+        val search = searchQueryDao.findByQuery(searchQuery)
 
-        searchQueryDao.delete(search)
+        if (search != null)
+            searchQueryDao.delete(search)
+
         getSearchHistory()
     }
